refactor: rename routes and update metadata
- Rename /about → /buyer-information (canonical URL)
- Rename /contact → /select-products (canonical URL)
- Keep /about and /contact as client-side redirects for legacy support
- Update homepage CTA link to /buyer-information
- Add LOGO NO BG.png as favicon in layout metadata
- Both new routes preserve sessionStorage handoff for buyer data flow

New routes:
- /buyer-information: buyer information form → redirects to /select-products
- /select-products: product selection and order confirmation
- /about: redirects to /buyer-information (legacy support)
- /contact: redirects to /select-products (legacy support)
